refactor: combat phase-2 slice 10 — part2/prob tables and morale helper (#3865)
Migrate probabilistic, part2, and spy/civilian resolver suites to
table-driven scenario loops; extract combatSideMoraleMultiplier for land
auto-resolve; combat test LOC 2623 → 1921.

Refs #3865
